requirements.txt
# Core data science and ML libraries
pandas>=1.5.0
numpy>=1.21.0
scikit-learn>=1.2.0
matplotlib>=3.5.0
seaborn>=0.11.0

# API and data fetching
requests>=2.28.0
yfinance>=0.1.87

# Technical analysis
ta>=0.10.0

# Time series and econometrics
arch>=5.3.0

# Google Trends
pytrends>=4.8.0

# Model persistence
joblib>=1.1.0

# Environment and utilities
python-dotenv>=0.19.0
joblib>=1.1.0

# Jupyter (for notebooks)
jupyter>=1.0.0

# Additional utilities
httpx>=0.24.0
